packages/one-context/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "one-context"
version = "0.1.0"
description = "one-context — multi-repository AI workspace control plane (CLI: onecxt)"
license = "MIT"
license-files = ["../../LICENSE"]
authors = [{name = "aiplaystation"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0.1",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Build Tools",
]

[project.urls]
Homepage = "https://github.com/harnessworld/one-context"
Repository = "https://github.com/harnessworld/one-context"
Issues = "https://github.com/harnessworld/one-context/issues"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[project.scripts]
onecxt = "one_context.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["one_context*"]